+++ src/main/java/Lumberj3ck/indicators/SimpleMovingAverageIndicator.java
@@ -3,9 +3,9 @@ package Lumberj3ck.indicators;
 import java.util.ArrayList;
 
 public class SimpleMovingAverageIndicator {
-    public static Integer calculate(ArrayList<Integer> prices) {
+    public static Integer calculate(ArrayList<Double> prices) {
         int priceSum = 0;
-        for (int cp : prices) {
+        for (double cp : prices) {
             priceSum += cp;
         }
         return priceSum / prices.size();
